requirements.txt
# Core dependencies for OpenAI Agents workflow
openai-agents>=0.4.2
openai>=2.6.1
pydantic>=2.12.3
pydantic-settings>=2.11.0
python-dotenv>=1.2.1

# FastAPI for REST API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Monitoring and logging
python-json-logger>=2.0.7
